new file lib.js created
