+++ package.json
@@ -2,6 +2,7 @@
   "dependencies": {
     "console-table-printer": "^2.12.0",
     "inquirer": "^8.2.6",
+    "prompt": "^1.3.0",
     "uniqid": "^5.4.0"
   }
 }
